docs: create file to hold listeners
